Work towards python3 compatible use of StringIO
Add in try/except to catch import errors with cStringIO and use StringIO
from the io module instead. Also ensure that where write access is
needed, create the StringIO without an initial string.

Need to avoid passing an initial buffer value to cStringIO.StringIO() as
it will return a read-only StringI object. Instead create the object
first with no initial arguments to get a proper StringIO instance and
then write the desired buffer to it.