Fix mkdir race condition in LooseObjectDB.store
This replaces the conditional call to os.mkdir that raises an
unintended FileExistsError if the directory is created between the
check and the os.mkdir call, using a single os.makedirs call
instead, with exist_ok=True.

This way, we attempt creation in a way that produces no error if
the directory is already present, while still raising
FileExistsError if a non-directory filesystem entry (such as a
regular file) is present where we want the directory to be.