Fix top-of-file license URLs here in gitdb too
This is the gitdb part of the fix for the top-of-file license URLs
that have come to point to a page about a related but different
license from the one GitPython and gitdb are (intended to be)
offered under.

See https://github.com/gitpython-developers/GitPython/pull/1662
for details about the problem and how it came about.