No longer treat 3.12 as experimental on CI
Since Python 3.12.0 stable has been released, as well as now being
available via setup-python, per:
https://github.com/actions/python-versions/blob/main/versions-manifest.json

The main practical effect of this is that continue-on-error is no
longer set to true for 3.12, so 3.12 is no longer special-cased to
refrain from cancelling other test jobs when its test job fails.

Another effect is that 3.12 can longer be selected as a prerelease.