Never add a vendored smmap directory to sys.path
This removes the logic that appended the git submodule directory
for smmap to sys.path under most circumstances when the version
of gitdb was not from PyPI. Now gitdb does not modify sys.path.

See https://github.com/gitpython-developers/GitPython/issues/1717
and https://github.com/gitpython-developers/GitPython/pull/1720 for
context. This change is roughly equivalent to the change to
GitPython, though as noted the behavior being eliminated is subtly
different here and there.