Improve description of backoff sequence in db.loose
The sequence of backoff wait times used in `gitdb.db.loose` is
quadratic rather than exponential, as discussed in:
https://github.com/gitpython-developers/gitdb/pull/115#discussion_r1903215598

This corrects the variable name by making it more general, and the
comment by having it explicitly describe the backoff as quadratic.

This is conceptually related to GitoxideLabs/gitoxide#1815, but
this is a non-breaking change, as no interfaces are affected: only
a local variable and comment.