Add CI test job for no-GIL ("threaded") Python 3.13
See https://github.com/gitpython-developers/GitPython/issues/2005.

The rationale is that, while this is probably less important to do
in gitdb and smmap, any failure that arises for this in GitPython
would likely raise the question of whether a correspond problem has
begun to occur in gitdb and smmap. (Both gitdb and smmap provide
helpers used in GitPython even when the in-memory object database
is not used, and failures may plausibly occur for reasons other
than code changes because of the finicky nature of concurrency bugs
and the potential for interactions affected by the runner image.)